refactor: 좋아요 기능 구현

+++ controllers/like.js
@@ -1,4 +1,5 @@
 import * as likeRepository from '../data/like.js';
+import * as waffleCardRepository from '../data/waffleCard.js';
 import { isValidId } from './utils.js';
 
 export const createLike = async (req, res) => {
@@ -11,7 +12,22 @@ export const createLike = async (req, res) => {
     });
   }
 
+  const likes = await likeRepository.getAllByUserId(userId);
+  for (let i = 0; i < likes.length; i++) {
+    if (likes[i].waffleCardId === waffleCardId) {
+      return res.status(400).json({
+        message: '이미 좋아요를 누른 와플 카드입니다.',
+      });
+    }
+  }
+
+  const waffleCard = await waffleCardRepository.getById(waffleCardId);
+  if (!waffleCard) {
+    return res.status(404).json({ message: '존재하지 않는 와플카드입니다.' });
+  }
+
   const like = await likeRepository.create(userId, waffleCardId);
+  await waffleCardRepository.countUpLike(waffleCardId);
   res.status(201).json({ id: like.id, userId, waffleCardId });
 };
 
@@ -25,6 +41,20 @@ export const deleteLike = async (req, res) => {
     });
   }
 
+  const likes = await likeRepository.getAllByUserId(userId);
+  if (!likes.map(like => like.userId).includes(userId)) {
+    return res.status(400).json({
+      message: '좋아요를 누른 와플 카드가 아닙니다.',
+    });
+  }
+
+  const waffleCard = await waffleCardRepository.getById(waffleCardId);
+
+  if (!waffleCard) {
+    return res.status(404).json({ message: '존재하지 않는 와플카드입니다.' });
+  }
+
   await likeRepository.remove(userId, waffleCardId);
+  await waffleCardRepository.countDownLike(waffleCardId);
   res.sendStatus(204);
 };
